rework timer with Dongting
